2024/hackceler8/server/requirements.txt
numpy >= 1.25.2
xxhash >= 3.3.0
dill == 0.3.7
pylint-venv >= 3.0.0
pylint >= 3.0.0
imgui >= 2.0.0
pillow~=10.4.0
pyrr~=0.10.3
# 5.11 breaks renderdoc
moderngl == 5.10.0
# https://github.com/moderngl/moderngl-window/pull/195, https://github.com/moderngl/moderngl-window/pull/194
-e git+https://github.com/implr/moderngl-window.git@d089e9b1499449f59d0b138e8501cbeae50844bd#egg=moderngl_window
